Finish spliting examples
