+++ examples/reuleaux_polygon.rs
@@ -25,6 +25,6 @@ fn main() {
 fn write_example(shape: &CSG, name: &str) {
     let _ = fs::write(
         Path::new(PATH).join(name).with_extension("stl"),
-        shape.to_stl_ascii(name),
+        shape.to_stl_binary(name).unwrap(),
     );
 }
